+++ .storybook/stories.js
@@ -25,7 +25,8 @@ import '../stories/Page/index.story.js'; // 2.5 Page
 import '../stories/PageHeader/index.story.js'; // 2.5 + PageHeader
 import '../stories/Page/FullPageExample.js'; // 2.5 + Page Example
 import '../stories/StatsWidget'; // 2.6 StatsWidget
-import '../stories/GridWithCardLayout'; // Grid & Card
+import '../stories/Grid';
+import '../stories/Card';
 
 
 // 3. Inputs