+++ src/Grid/Grid.js
@@ -4,38 +4,25 @@ import Card from '../Card';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 
-class RawContainer extends Component {
+const childrenProp = {children: PropTypes.node};
 
-  static propTypes = {
-    children: PropTypes.node
-  };
+const RawContainer = ({children}) =>
+  <div
+    className={styles.rawContainer}
+    children={children}
+    />;
 
-  render() {
-    return (
-      <div className={styles['raw-container']}>
-        {this.props.children}
-      </div>
-    );
-  }
-}
+RawContainer.propTypes = childrenProp;
 
-class Container extends Component {
+const Container = ({children}) =>
+  <div
+    className={styles.wixContainer}
+    children={children}
+    />;
 
-  static propTypes = {
-    children: PropTypes.node
-  };
-
-  render() {
-    return (
-      <div className={styles.wixContainer}>
-        {this.props.children}
-      </div>
-    );
-  }
-}
+Container.propTypes = childrenProp;
 
 class Columns extends Component {
-
   static propTypes = {
     children: PropTypes.node,
     className: PropTypes.string,
@@ -49,44 +36,54 @@ class Columns extends Component {
   };
 
   render() {
+    const {className, rtl, stretchViewsVertically, dataHook, children} = this.props;
+
     const rowClasses = classNames(
       styles.row,
-      this.props.className,
+      className,
       {
-        [styles.rtl]: this.props.rtl,
-        [styles.stretch_vertically_row]: this.props.stretchViewsVertically
-      });
+        [styles.rtl]: rtl,
+        [styles.stretch_vertically_row]: stretchViewsVertically
+      }
+    );
 
     return (
-      <div className={rowClasses} data-hook={this.props.dataHook}>
-        {this.props.children}
-      </div>
+      <div
+        className={rowClasses}
+        data-hook={dataHook}
+        children={children}
+        />
     );
   }
 }
 
 class AutoAdjustedColumns extends Component {
-
   DEFAULT_MAX_SPAN = 12;
+
   static propTypes = {
     children: PropTypes.node
   };
 
   render() {
-    const cssClasses = classNames(styles.row, styles.flexContainer);
     const children = this.props.children;
     const cols = Array.isArray(children) ? children : [children];
     const spanSize = Math.floor(this.DEFAULT_MAX_SPAN / cols.length);
+
     return (
-      <div className={cssClasses}>
-        {cols.map((child, index) => <Col span={spanSize} key={index}>{child}</Col>)}
+      <div className={classNames(styles.row, styles.flexContainer)}>
+        { cols.map((col, index) =>
+          <Col
+            span={spanSize}
+            key={index}
+            children={col}
+            />
+        ) }
       </div>
     );
   }
 }
 
 class Col extends Component {
-
   static propTypes = {
     children: PropTypes.node,
     className: PropTypes.string,
@@ -117,32 +114,54 @@ class Col extends Component {
   }
 
   render() {
+    const {
+      children,
+      className,
+      span,
+      rtl,
+      xs,
+      sm,
+      md,
+      lg,
+      xl,
+      dataHook
+    } = this.props;
+
     const columnClasses = classNames(
-      this.props.className,
+      className,
       styles.column,
-      {[styles.rtl]: this.props.rtl},
-      {[styles[`colXs${this.props.span}`]]: this.isLegalCol(this.props.span)},
-      {[styles[`colXs${this.props.xs}`]]: this.isLegalCol(this.props.xs)},
-      {[styles[`colSm${this.props.sm}`]]: this.isLegalCol(this.props.sm)},
-      {[styles[`colMd${this.props.md}`]]: this.isLegalCol(this.props.md)},
-      {[styles[`colLg${this.props.lg}`]]: this.isLegalCol(this.props.lg)},
-      {[styles[`colXl${this.props.xl}`]]: this.isLegalCol(this.props.xl)},
-      {[styles[`${this.props.xs}Xs`]]: this.isVisibleHidden(this.props.xs)},
-      {[styles[`${this.props.sm}Sm`]]: this.isVisibleHidden(this.props.sm)},
-      {[styles[`${this.props.md}Md`]]: this.isVisibleHidden(this.props.md)},
-      {[styles[`${this.props.lg}Lg`]]: this.isVisibleHidden(this.props.lg)},
+      {
+        [styles.rtl]: rtl,
+        [styles[`colXs${span}`]]: this.isLegalCol(span),
+        [styles[`colXs${xs}`]]: this.isLegalCol(xs),
+        [styles[`colSm${sm}`]]: this.isLegalCol(sm),
+        [styles[`colMd${md}`]]: this.isLegalCol(md),
+        [styles[`colLg${lg}`]]: this.isLegalCol(lg),
+        [styles[`colXl${xl}`]]: this.isLegalCol(xl),
+        [styles[`${xs}Xs`]]: this.isVisibleHidden(xs),
+        [styles[`${sm}Sm`]]: this.isVisibleHidden(sm),
+        [styles[`${md}Md`]]: this.isVisibleHidden(md),
+        [styles[`${lg}Lg`]]: this.isVisibleHidden(lg)
+      }
     );
+
     return (
-      <div className={columnClasses} data-hook={this.props.dataHook}>
-        {this.props.children}
-      </div>
+      <div
+        className={columnClasses}
+        data-hook={dataHook}
+        children={children}
+        />
     );
   }
 }
 
 export {
-  Container, RawContainer,
-  Columns, Columns as Row,
-  AutoAdjustedColumns, AutoAdjustedColumns as AutoAdjustedRow,
-  Col, Card
+  Container,
+  RawContainer,
+  Columns,
+  Columns as Row,
+  AutoAdjustedColumns,
+  AutoAdjustedColumns as AutoAdjustedRow,
+  Col,
+  Card
 };