+++ src/Input/Input.puppeteer.driver.js
@@ -2,7 +2,10 @@ const inputDriverFactory = async (component, page) => {
   const input = await component.$('input');
   return {
     element: () => component,
-    enterText: text => input.type(text),
+    enterText: async text => {
+      await input.focus();
+      await input.type(text);
+    },
     getText: () => page.evaluate(_input => _input.value, input)
   };
 };